Merged PR 16311: [Repair Item] Return a 400 HTTP status code for invalid ACLs instead of 500
## Describe your changes
-------------------------------------
Change details:
<!-- Please describe the current behavior that you are modifying, 'or' link to a relevant issue. -->
This PR addresses the repair item #47900.

For invalid ACLs, we would want to return 400s instead of 500s to the user. Currently, the patch records API would return 500s if the ACLs were missing an "@" symbol or if the ACLs were null.

ICM: https://portal.microsofticm.com/imp/v5/incidents/details/557901425/summary

## C1B testing

### Before these PR changes:
![image.png](https://dev.azure.com/OpenEnergyPlatform/dcd9d9fe-a1ca-4b31-bb47-0364ffe4fca7/_apis/git/repositories/a1db38cb-e726-44da-987b-7c15620ac1be/pullRequests/16311/attachments/image.png)

In the screenshot below, the highlighted part indicates that there's a missing "@" symbol between `owners` and `opendes`. For instance, a valid format of the ACL in this case would be `data.default.owners@opendes.test.com`

![image (12).png](https://dev.azure.com/OpenEnergyPlatform/dcd9d9fe-a1ca-4b31-bb47-0364ffe4fca7/_apis/git/repositories/a1db38cb-e726-44da-987b-7c15620ac1be/pullRequests/16311/attachments/image%20%2812%29.png)

After these changes, the returned HTTP status code is 400 when invalid ACLs are passed.

![image (11).png](https://dev.azure.com/OpenEnergyPlatform/dcd9d9fe-a1ca-4b31-bb47-0364ffe4fca7/_apis/git/repositories/a1db38cb-e726-44da-987b-7c15620ac1be/pullRequests/16311/attachments/image%20%2811%29.png)

![image (10).png](https://dev.azure.com/OpenEnergyPlatform/dcd9d9fe-a1ca-4b31-bb47-0364ffe4fca7/_apis/git/repositories/a1db38cb-e726-44da-987b-7c15620ac1be/pullRequests/16311/attachments/image%20%2810%29.png)

Using the ACL in the ICM https://portal.microsofticm.com/imp/v5/incidents/details/557901425/summary, a 400 status code is returned instead of a 500.

![image (13).png](https://dev.azure.com/OpenEnergyPlatform/dcd9d9fe-a1ca-4b31-bb47-0364ffe4fca7/_apis/git/repositories/a1db38cb-e726-44da-987b-7c15620ac1be/pullRequests/16311/attachments/image%20%2813%29.png)

## PR Checklist
-------------------------------------
* [ ] I have parameterized Cosmos DB queries in my code(if any) and confirm adherence to the [guidelines](https://dev.azure.com/OpenEnergyPlatform/Open%20Energy%20Platform/_wiki/wikis/ADME%20Engineering%20Docs/15562/Cosmos-db-Paramterized-queries).

One of the two following selections must be made. If the change is not merged back to OSDU Master, valid justification must be provided
* [x] I have created a MR into OSDU Master branch with my changes. Link to MR: https://community.opengroup.org/osdu/platform/system/storage/-/merge_requests/1061
* [ ] Change is not able to be merged to OSDU Master. I have documented in PR code and description why it is not able to be merged back into OSDU Master:

Related work items: #47900